requirements.txt
asgiref==3.3.4
astroid==2.5.6
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
dj-places==4.0.0
Django==3.2.3
django-bootstrap-v5==1.0.4
django-filer==2.0.2
django-js-asset==1.2.2
django-libs==2.0.3
django-mptt==0.12.0
django-polymorphic==3.0.0
django-scheduler==0.9.5
easy-thumbnails==2.7.1
haversine==2.3.1
icalendar==4.0.7
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.20.3
pandas==1.2.4
pgeocode==0.3.0
Pillow==8.2.0
PyJWT==1.7.1
pylint==2.8.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
twilio==6.59.1
Unidecode==1.1.2
urllib3==1.26.5
wrapt==1.12.1
gunicorn